requirements.txt
langchain
langchain-community
langchain-groq
faiss-cpu
pymupdf
sentence-transformers
streamlit
pyyaml
